requirements.txt
fastapi==0.101.1  # latest version compatible with Pydantic 2.x
pydantic>=2.0.0,<3.0.0
pydantic[email]>=2.0.0
google-genai==1.6.0
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy>=2.0.0
aiosqlite==0.18.0
python-json-logger==2.0.7

